requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic==2.11.2
asyncio==3.4.3
mcp==1.6.0
Flask==3.1.0